Upload tests terminal logs to Browserstack and fixed incomplete consolidated logs issue

+++ conftest.py
@@ -52,6 +52,24 @@ def test_obj(base_url, browser, browser_version, os_version, os_name, remote_fla
                 test_obj.execute_javascript("lambda-status=failed")
                 test_obj.teardown()
 
+        elif os.getenv('REMOTE_BROWSER_PLATFORM') == 'BS' and remote_flag.lower() == 'y':
+            response = upload_test_logs_to_browserstack(test_obj.log_name,test_obj.session_url)
+            if isinstance(response, dict) and "error" in response:
+                # Handle the error response returned as a dictionary
+                test_obj.write(f"Error: {response['error']}",level='error')
+                if "details" in response:
+                    test_obj.write(f"Details: {response['details']}",level='error')
+                    test_obj.write("Failed to upload log file to BrowserStack",level='error')
+            else:
+                # Handle the case where the response is assumed to be a response object
+                if response.status_code == 200:
+                    test_obj.write("Log file uploaded to BrowserStack session successfully.")
+                else:
+                    test_obj.write(f"Failed to upload log file. Status code: {response.status_code}",level='error')
+                    test_obj.write(response.text,level='error')
+
+            test_obj.teardown()
+
         else:
             test_obj.wait(3)
             test_obj.teardown()
@@ -91,6 +109,22 @@ def test_mobile_obj(mobile_os_name, mobile_os_version, device_name, app_package,
 
         yield test_mobile_obj
 
+        if os.getenv('REMOTE_BROWSER_PLATFORM') == 'BS' and remote_flag.lower() == 'y':
+            response = upload_test_logs_to_browserstack(test_mobile_obj.log_name,test_mobile_obj.session_url,appium_test = True)
+            if isinstance(response, dict) and "error" in response:
+                # Handle the error response returned as a dictionary
+                test_obj.write(f"Error: {response['error']}",level='error')
+                if "details" in response:
+                    test_obj.write(f"Details: {response['details']}",level='error')
+                    test_obj.write("Failed to upload log file to BrowserStack",level='error')
+            else:
+                # Handle the case where the response is assumed to be a response object
+                if response.status_code == 200:
+                    test_mobile_obj.write("Log file uploaded to BrowserStack session successfully.")
+                else:
+                    test_mobile_obj.write(f"Failed to upload log file. Status code: {response.status_code}",level='error')
+                    test_mobile_obj.write(response.text,level='error')
+
         #Teardown
         test_mobile_obj.wait(3)
         test_mobile_obj.teardown()
@@ -120,6 +154,44 @@ def test_api_obj(request, interactivemode_flag, api_url=base_url_conf.api_base_u
         print("Exception when trying to run test:%s" % __file__)
         print("Python says:%s" % str(e))
 
+def upload_test_logs_to_browserstack(log_name, session_url, appium_test = False):
+    "Upload log file to provided BrowserStack session"
+    try:
+        from integrations.cross_browsers.BrowserStack_Library import BrowserStack_Library # pylint: disable=import-error,import-outside-toplevel
+        # Initialize BrowserStack object
+        browserstack_obj = BrowserStack_Library()
+
+        # Build log file path
+        log_file_dir = os.path.abspath(os.path.join(os.path.dirname(__file__),'log'))
+        log_file = log_file_dir + os.sep + 'temp_' + log_name
+
+        # Check if the log file exists
+        if not os.path.isfile(log_file):
+            raise FileNotFoundError(f"Log file '{log_file}' not found.")
+
+        # Extract session ID from the provided session URL
+        session_id = browserstack_obj.extract_session_id(session_url)
+        if not session_id:
+            raise ValueError(f"Invalid session URL provided: '{session_url}'")
+
+        # Upload the log file to BrowserStack
+        response = browserstack_obj.upload_terminal_logs(log_file,session_id,appium_test)
+
+        return response
+
+    except ImportError as e:
+        return {"error": "Failed to import BrowserStack_Library.", "details": str(e)}
+
+    except FileNotFoundError as e:
+        return {"error": "Log file not found.", "details": str(e)}
+
+    except ValueError as e:
+        return {"error": "Invalid session URL.", "details": str(e)}
+
+    except Exception as e:
+        # Handle any other unexpected exceptions
+        return {"error": "An unexpected error occurred while uploading logs to BrowserStack.", "details": str(e)}
+
 @pytest.fixture
 def testname(request):
     "pytest fixture for testname"
@@ -137,7 +209,7 @@ def testname(request):
 def browser(request):
     "pytest fixture for browser"
     try:
-       return request.config.getoption("--browser")
+        return request.config.getoption("--browser")
 
     except Exception as e:
         print("Exception when trying to run test: %s"%__file__)
@@ -418,8 +490,8 @@ def reportportal_service(request):
     "pytest service fixture for reportportal"
     reportportal_pytest_service = None
     try:
-       if request.config.getoption("--reportportal"):
-           reportportal_pytest_service = request.node.config.py_test_service
+        if request.config.getoption("--reportportal"):
+            reportportal_pytest_service = request.node.config.py_test_service
     except Exception as e:
         print("Exception when trying to run test: %s"%__file__)
         print("Python says:%s"%str(e))
@@ -452,23 +524,24 @@ def pytest_sessionstart(session):
     Perform cleanup at the start of the test session.
     Delete the consolidated log file and temporary log files if present.
     """
-    source_directory = "log"
-    log_file_name = "*.log-temp"
-    consolidated_log_file = os.path.join(source_directory, "consolidated_log.txt")
+    if not hasattr(session.config, "workerinput"):  # Executes during the main session only
+        source_directory = "log"
+        log_file_name = "temp_*.log"
+        consolidated_log_file = os.path.join(source_directory, "consolidated_log.txt")
 
-    # Delete the consolidated log file
-    if os.path.exists(consolidated_log_file):
-        try:
-            os.remove(consolidated_log_file)
-        except OSError as error:
-            print(f"Error removing existing consolidated log file: {error}")
+        # Delete the consolidated log file
+        if os.path.exists(consolidated_log_file):
+            try:
+                os.remove(consolidated_log_file)
+            except OSError as error:
+                print(f"Error removing existing consolidated log file: {error}")
 
-    # Delete all temporary log files if present
-    for temp_log_file in glob.glob(os.path.join(source_directory, log_file_name)):
-        try:
-            os.remove(temp_log_file)
-        except OSError as error:
-            print(f"Error removing temporary log file: {error}")
+        # Delete all temporary log files if present
+        for temp_log_file in glob.glob(os.path.join(source_directory, log_file_name)):
+            try:
+                os.remove(temp_log_file)
+            except OSError as error:
+                print(f"Error removing temporary log file: {error}")
 
 def pytest_sessionfinish(session, exitstatus):
     """
@@ -476,30 +549,30 @@ def pytest_sessionfinish(session, exitstatus):
     The temporary log files are consolidated into a single log file 
     and later deleted.
     """
-    source_directory = "log"
-    log_file_name = "*.log-temp"
+    if not hasattr(session.config, "workerinput"):  # Executes during the main session only
+        source_directory = "log"
+        log_file_name = "temp_*.log"
+        consolidated_log_file = os.path.join(source_directory, "consolidated_log.txt")
 
-    consolidated_log_file = os.path.join(source_directory, "consolidated_log.txt")
+        #Detach all handlers from the logger inorder to release the file handle
+        #which can be used for deleting the temp files later
+        logger.remove(None)
 
-    #Detach all handlers from the logger inorder to release the file handle
-    #which can be used for deleting the temp files later
-    logger.remove(None)
-
-    #Consolidate the temporary log files into the consolidated log file
-    try:
-        with open(consolidated_log_file, "a") as final_log:
-            for file_name in glob.glob(os.path.join(source_directory, log_file_name)):
-                source_file = None
-                try:
-                    with open(file_name, "r") as source_file:
-                        shutil.copyfileobj(source_file, final_log)
-                    os.remove(file_name)
-                except FileNotFoundError as error:
-                    print(f"Temporary log file not found: {error}")
-                except Exception as error:
-                    print(f"Error processing the temporary log file: {error}")
-    except OSError as error:
-        print(f"Error processing consolidated log file: {error}")
+        #Consolidate the temporary log files into the consolidated log file
+        try:
+            with open(consolidated_log_file, "a") as final_log:
+                for file_name in glob.glob(os.path.join(source_directory, log_file_name)):
+                    source_file = None
+                    try:
+                        with open(file_name, "r") as source_file:
+                            shutil.copyfileobj(source_file, final_log)
+                        os.remove(file_name)
+                    except FileNotFoundError as error:
+                        print(f"Temporary log file not found: {error}")
+                    except Exception as error:
+                        print(f"Error processing the temporary log file: {error}")
+        except OSError as error:
+            print(f"Error processing consolidated log file: {error}")
 
 @pytest.hookimpl()
 def pytest_configure(config):
